+++ TestingAndCalibrationLabs/TestingAndCalibrationLabs.Business/Core/Model/Security/CustomClaimTypes.cs
@@ -1,0 +1,11 @@
+﻿namespace TestingAndCalibrationLabs.Business.Core.Model
+{
+    /// <summary>
+    /// Claims Types
+    /// </summary>
+
+    public class CustomClaimTypes
+    {
+        public const string Permission = "Application.Permission";
+    }
+}
