+++ TestingAndCalibrationLabs/TestingAndCalibrationLabs.Business/Core/Model/Security/LoginRequest.cs
@@ -1,6 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace TestingAndCalibrationLabs.Business.Core.Model
 {
