+++ TestingAndCalibrationLabs/TestingAndCalibrationLabs.Business/Core/Model/Security/PolicyTypes.cs
@@ -1,0 +1,15 @@
+﻿namespace TestingAndCalibrationLabs.Business.Core.Model
+{
+    /// <summary>
+    /// Policy Type User Permission
+    /// </summary>
+
+    public static class PolicyTypes
+    {
+        public static class Users
+        {
+            public const string Manage = "users.manage.policy";
+            public const string EditRole = "users.edit.role.policy";
+        }
+    }
+}
