+++ TestingAndCalibrationLabs/TestingAndCalibrationLabs.Web.UI/Controllers/SecurityController.cs
@@ -1,0 +1,59 @@
+﻿using AutoMapper;
+using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.Rendering;
+using System.Collections.Generic;
+using System.Linq;
+using TestingAndCalibrationLabs.Business.Common;
+using TestingAndCalibrationLabs.Business.Core.Interfaces;
+using TestingAndCalibrationLabs.Business.Core.Model;
+using TestingAndCalibrationLabs.Web.UI.Models;
+
+namespace TestingAndCalibrationLabs.Web.UI.Controllers
+{
+    public class SecurityController : Controller
+    {
+        private readonly IAuthenticationService _authenticationService;
+        private readonly IOrganizationService _orgnizationService;
+        private readonly IMapper _mapper;
+        public SecurityController(IAuthenticationService authenticationService, IOrganizationService orgnizationService, IMapper mapper)
+        {
+            _authenticationService = authenticationService;
+            _orgnizationService = orgnizationService;
+            _mapper = mapper;
+        }
+
+        /// <summary>
+        /// UI Shows the Orgnizations names in dropdown list
+        /// </summary>
+        /// <returns></returns>
+        [HttpGet]
+        public IActionResult Login()
+        {
+            List<Business.Core.Model.Organization> organizations = _orgnizationService.Get();
+            List<SelectListItem> organizationNames = organizations.Select(x => new SelectListItem { Text = x.OrgName, Value = x.Id.ToString() }).ToList();
+            ViewBag.Organizations = organizationNames;
+            return View();
+        }
+
+        /// <summary>
+        /// Method to get the Login details from UI and Process Login.
+        /// </summary>
+        /// <returns></returns>
+
+        [HttpPost]
+        public IActionResult Login(LoginDTO loginRequest)
+        {
+            var loginReq = new LoginRequest { UserName = loginRequest.UserName, Password = loginRequest.Password };
+            RequestResult<LoginToken> result = _authenticationService.Login(loginReq);
+
+            if (result.IsSuccessful)
+            {
+                HttpContext.Session.SetString("Token", result.RequestedObject.AccessToken);
+
+                return Json(new { status = true, message = "Login Successfull!" });
+            }
+            return View();
+        }
+    }
+}
