+++ TestingAndCalibrationLabs/TestingAndCalibrationLabs.Web.UI/Controllers/UserController.cs
@@ -31,7 +31,7 @@ namespace TestingAndCalibrationLabs.Web.UI.Controllers
         /// </summary>
         /// <returns></returns>
         [HttpPost]
-        public IActionResult UserRegistration(UserDTO userRequest)
+        public IActionResult Add(UserDTO userRequest)
         {
             var userModel = _mapper.Map<UserDTO,User>(userRequest);
             var result = _authenticationService.Add(userModel, userRequest.Password);
