+++ TestingAndCalibrationLabs/TestingAndCalibrationLabs.Web.UI/appsettings.json
@@ -7,7 +7,7 @@
     }
   },
   "ConnectionStrings": {
-    "IdentityDBLocal": "Server=localhost;Database=TestingAndCalibrationLabs; User ID=sa;Password=admin;"
+    "IdentityDBLocal": "Server=localhost;Database=SdtDbDev; User ID=sa;Password=Saylo@13;"
   },
 
   "Smtp": {
